fix warning when comparing int and QVariant::type
[git-p4: depot-paths = "//depot/qtjambi/main/": change = 281088]

+++ qtjambi/qtjambi_core.cpp
@@ -302,7 +302,7 @@ QVariant qtjambi_to_qvariant(JNIEnv *env, jobject java_object)
     QString className = manager.getInternalTypeName(fullName, QtJambiTypeManager::ArgumentType);
     int type = !className.isEmpty()
         ? manager.metaTypeOfInternal(className, QtJambiTypeManager::ArgumentType)
-        : QVariant::Invalid;
+        : (int)QVariant::Invalid;
 
     void *copy = 0;
     bool destroyCopy = false;
